菜单缓存优化

+++ dayrui/Fcms/Model/Menu.php
@@ -177,7 +177,7 @@ class Menu extends \Phpcmf\Model {
         }
 
         // 内容模块入库用户菜单
-        if ($config['system'] == 1) {
+        if ($config['system'] == 1 && dr_is_app('member')) {
             $left = $this->db->table('member_menu')->where('mark', 'content-module')->get()->getRowArray();
             if ($left) {
                 // 查询模块菜单
@@ -319,7 +319,7 @@ class Menu extends \Phpcmf\Model {
             }
         }
 
-        if ($menu['member']) {
+        if ($menu['member'] && dr_is_app('member')) {
             // 用户菜单
             foreach ($menu['member'] as $mark => $top) {
                 // 插入顶级菜单
@@ -389,7 +389,7 @@ class Menu extends \Phpcmf\Model {
             }
         }
 
-        if ($menu['member']) {
+        if ($menu['member'] && dr_is_app('member')) {
             // 用户菜单
             foreach ($menu['member'] as $mark => $top) {
                 // 插入顶级菜单
@@ -431,11 +431,11 @@ class Menu extends \Phpcmf\Model {
 
 
         $this->db->table('admin_menu')->where('mark', 'app-'.$dir)->delete();
-        $this->db->table('member_menu')->where('mark', 'app-'.$dir)->delete();
+        dr_is_app('member') && $this->db->table('member_menu')->where('mark', 'app-'.$dir)->delete();
         $this->db->table('admin_min_menu')->where('mark', 'app-'.$dir)->delete();
 
         $this->db->table('admin_menu')->like('mark', 'app-'.$dir.'%')->delete();
-        $this->db->table('member_menu')->like('mark', 'app-'.$dir.'%')->delete();
+        dr_is_app('member') && $this->db->table('member_menu')->like('mark', 'app-'.$dir.'%')->delete();
         $this->db->table('admin_min_menu')->like('mark', 'app-'.$dir.'%')->delete();
     }
 
@@ -511,7 +511,7 @@ class Menu extends \Phpcmf\Model {
                 }
             }
 
-        } elseif ($table == 'member') {
+        } elseif (dr_is_app('member') && $table == 'member') {
             // 清空表
             $this->db->table('member_menu')->truncate();
             $this->db->table('member_menu')->emptyTable();
@@ -723,36 +723,39 @@ class Menu extends \Phpcmf\Model {
             $menu['admin'] = $list;
         }
         // member 菜单
-        $data = $this->db->table('member_menu')->where('hidden', 0)->orderBy('displayorder ASC,id ASC')->get()->getResultArray();
-        if ($data) {
-            $list = [
-                'url' => [],
-                'uri' => [],
-            ];
-            foreach ($data as $t) {
-                if ($t['pid'] == 0) {
-                    $t['site'] = dr_string2array($t['site']);
-                    $t['group'] = dr_string2array($t['group']);
-                    $list['url'][$t['id']] = $t;
-                    foreach ($data as $n) {
-                        $n['site'] = dr_string2array($n['site']);
-                        $n['group'] = dr_string2array($n['group']);
-                        if ($n['pid'] == $t['id']) {
-                            $list['url'][$t['id']]['link'][$n['id']] = $n;
-                            $n['uri'] && $list['uri'][$n['uri']] = [
-                                'id' => $n['id'],
-                                'pid' => $t['id'],
-                                'icon' => $n['icon'],
-                                'picon' => $t['icon'],
-                                'name' => $n['name'],
-                                'pname' => $t['name'],
-                            ];
+        if (dr_is_app('member')) {
+            $data = $this->db->table('member_menu')->where('hidden', 0)->orderBy('displayorder ASC,id ASC')->get()->getResultArray();
+            if ($data) {
+                $list = [
+                    'url' => [],
+                    'uri' => [],
+                ];
+                foreach ($data as $t) {
+                    if ($t['pid'] == 0) {
+                        $t['site'] = dr_string2array($t['site']);
+                        $t['group'] = dr_string2array($t['group']);
+                        $list['url'][$t['id']] = $t;
+                        foreach ($data as $n) {
+                            $n['site'] = dr_string2array($n['site']);
+                            $n['group'] = dr_string2array($n['group']);
+                            if ($n['pid'] == $t['id']) {
+                                $list['url'][$t['id']]['link'][$n['id']] = $n;
+                                $n['uri'] && $list['uri'][$n['uri']] = [
+                                    'id' => $n['id'],
+                                    'pid' => $t['id'],
+                                    'icon' => $n['icon'],
+                                    'picon' => $t['icon'],
+                                    'name' => $n['name'],
+                                    'pname' => $t['name'],
+                                ];
+                            }
                         }
                     }
                 }
+                $menu['member'] = $list;
             }
-            $menu['member'] = $list;
         }
+
         // admin-min 菜单
         $data = $this->db->table('admin_min_menu')->where('hidden', 0)->orderBy('displayorder ASC,id ASC')->get()->getResultArray();
         if ($data) {
